+++ RELEASES.md
@@ -5,6 +5,7 @@ Released on TBD (UTC).
 ### Breaking Changes
 - Removed `OrderBookSnapshot` (redundant as can be represented as an initial CLEAR followed by deltas)
 - Removed `OrderBookData` (redundant)
+- Renamed `inverse_as_quote` to `use_quote_for_inverse` (ambiguous name, only applicable for notional cals on inverse instruments)
 - Changed `Data` contract (custom data), [see docs](https://docs.nautilustrader.io/develop/concepts/advanced/data.html)
 
 ### Enhancements